pyproject.toml
[tool.poetry]
name = "bsport"
version = "0.1.0"
description = ""
authors = ["Thomas Berdy <thomas.berdy@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Django = "~3.1.5"
django-extensions = "~3.1.0"
djangorestframework = "~3.12.2"
gunicorn = "~20.0.4"
psycopg2 = "~2.8.6"
pynamodb = "~4.3.3"
uvicorn = {extras = ["standard"], version = "~0.13.3"}

[tool.poetry.dev-dependencies]
ipdb = "*"
ipython = "*"
jedi = "==0.17.2"  # ipython bug
factory-boy = "~3.2.0"
pytest = "~6.2.1"
pytest-black = "~0.3.12"
pytest-django = "~4.1.0"
pytest-cov = "~2.10.1"
pytest-env = "~0.6.2"
pytest-sugar = "~0.9.4"

[tool.semantic_release]
branch = "master"
version_variable = "pyproject.toml:version"
version_source = "commit"
hvcs = "gitlab"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
